% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisprep.R
\name{actFun}
\alias{actFun}
\title{Animal Diel Activity}
\usage{
actFun(
  x,
  split = NULL,
  include,
  return = "species",
  bw = NULL,
  rep = 999,
  pp = F,
  cores.left = NULL
)
}
\arguments{
\item{x}{data.frame. A data frame produced by the \code{\link{calculateEvents}} function available in this package.}

\item{split}{String or NULL. How should the data be grouped before running \code{\link[activity]{fitact}} on the data. This can take NULL for no grouping or a vector of column names or index numbers. The default is NULL.}

\item{include}{String. Which species do you want to run diel activity on? Use form c("Species 1", "Species 2", "etc.").}

\item{return}{String. Should the species or the grouping factor be placed on top of the output list. Valid options include c("species", "group", "both"). This defaults to "species". If an inproper option is provided, the function outputs as "both".}

\item{bw}{Numeric or NULL. Do you want to specify a bandwidth for the kernel density estimation of the activity distribution? The default is NULL where the function with calculate bandwidth internally.}

\item{rep}{Numeric. The number of bootstraps that should be used to estimate the confidence intervals of the diel activity distribution.}

\item{pp}{Logical. Should this function take advantage of parallel processing.
The default is FALSE. Because the function separates tasks by camera directory, it can use parallel processing to run multiple cameras at the same time.
This is currently set up to run in Windows OS so I do not know if this will work on a Mac or Linux system.
If you want this functionality on a Unix device and know how to set it up, let me know and I can incorporate it.}

\item{cores.left}{Numeric. How many cores should be left available when using parallel processing?
The default is NULL. This is only necessary when pp=TRUE.
If left at the default, the function will default to 2 cores remaining which is generally enough to continue using a PC while the function runs.
I would set this to be greater than 0, otherwise the function will use the entire processing power of your computer.
To see how many cores you have available for parallel processing, use: parallel::detectCores().}
}
\value{
data:
These are the data tables for each species (and split) used to produce the activity distribution.

activity:
These are the estimations of diel activity for each species.
}
\description{
A function for calculating diel activity for selected species. The function can optionally also split your data based on a user-defined column.
}
\details{
This function now properly checks for and adjusts the confidence interval estimation for the number of observations used in the calculation.
This function will not calculate diel activity when there are less than 70 observations for a species-group and does provide an appropriate warning when this occurs.
See \code{\link[activity]{fitact}} for details on how confidence interval estimation should occur.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\references{
Rowcliffe, J. M., R. Kays, B. Kranstauber, C. Carbone, and P. A. Jansen. 2014. Quantifying levels of animal activity using camera trap data. Methods in Ecology and Evolution 5:1170-1179. 10.1111/2041-210x.12278

Frey, S., J. T. Fisher, A. C. Burton, and J. P. Volpe. 2017. Investigating animal activity patterns and temporal niche partitioning using camera-trap data: challenges and opportunities. Remote Sensing in Ecology and Conservation 3:123-132. 10.1002/rse2.60

Ramesh, T., R. Kalle, K. Sankar, and Q. Qureshi. 2012. Spatio-temporal partitioning among large carnivores in relation to major prey species in Western Ghats. Journal of Zoology 287:269-275. 10.1111/j.1469-7998.2012.00908.x
}
\seealso{
activity package: \url{https://cran.r-project.org/web/packages/activity/index.html}, \code{\link[activity]{activity}}

\code{\link{calculateEvents}}
}
\concept{camera trapping}
\concept{diel activity}
\keyword{manip}
\keyword{univar}
