% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisprep.R
\name{actPlot}
\alias{actPlot}
\title{Prepare diel activity plots}
\usage{
actPlot(act, top = "species")
}
\arguments{
\item{act}{list. The list object outputted by the \code{\link{actFun}} function. This must contain either species or the grouping factor as the top level of the list}

\item{top}{character. What was outputted as the top level list? This can be one of c("species", "group"). This should be the same as what is asked for in the "return" parameter of the \code{\link{actFun}} function.
It does not matter which is which, and fundamentally, it will not change anything about the output. It only affects the naming of two columns in the output.}
}
\value{
data.frame containing the probability density function information for x if it is centered on 12 noon (day) or 12 midnight (night) and y in radians, density, or frequency
}
\description{
A function for taking an actmod object and prepping it for plotting in ggplot
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\references{
Rowcliffe, J. M., R. Kays, B. Kranstauber, C. Carbone, and P. A. Jansen. 2014. Quantifying levels of animal activity using camera trap data. Methods in Ecology and Evolution 5:1170-1179. 10.1111/2041-210x.12278
}
\seealso{
activity package: \url{https://cran.r-project.org/web/packages/activity/index.html}, \code{\link[activity]{activity}}

activity package plotting code is viewable on github: \url{https://github.com/MarcusRowcliffe/activity}

\code{\link{actFun}}
}
\concept{diel activity}
\concept{plotting}
\keyword{dplot}
\keyword{manip}
