% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timelapse.R
\name{APFun_Timelapse}
\alias{APFun_Timelapse}
\title{Convert a Timelapse csv to a format for use with APFun_env}
\usage{
APFun_Timelapse(timelapse)
}
\arguments{
\item{timelapse}{data.frame. A data frame representing a properly formatted Timelapse csv file}
}
\value{
An R object formatted in the same style as a dataorganize text file.
}
\description{
This function converts a Timelapse csv file to the dataorganize file output format. This was done this way because many of my core functions for processing camera trap data depend on the existence of a data frame created by APFun_env. Converting a timelapse file to this format is just the easiest way to maintain consistency.
}
\note{
This function is designed to make timelapse files compatible with DataOrganize files which are used for most of the analyses in this package
}
\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{APFun_env}}
}
\concept{camera trapping}
\concept{timelapse}
\keyword{datagen}
