% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timelapse.R
\name{bestPics}
\alias{bestPics}
\title{Extract Best Pictures from Timelapse}
\usage{
bestPics(x, in.dir, out.dir, copy = T, sorted = F)
}
\arguments{
\item{x}{A data frame from a timelapse file}

\item{in.dir}{String. The directory containing the pictures. This can either be an unsorted directory or a sorted directory. Keep in mind, these assume my very specific directory format.}

\item{out.dir}{String. The directory where you want the best pictures to be stored. This can be anything.}

\item{copy}{Logical. Should the images be copied into the new directory?}

\item{sorted}{Logical. Is the in.dir an unsorted folder or a sorted folder?}
}
\value{
A data frame containing the file path info of the best pics.
}
\description{
This function uses a timelapse csv to extract photos tagged as a best picture and copies them to a specified location.
}
\examples{
\dontrun{
# No example provided
}
}
\concept{camera trapping}
\concept{timelapse}
\keyword{files}
\keyword{manip}
