% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{cameraDiagnostics}
\alias{cameraDiagnostics}
\title{Camera basic diagnostics}
\usage{
cameraDiagnostics(x, cam_dir_level, from_bottom)
}
\arguments{
\item{x}{data.frame. A data frame outputted from the \code{\link{cameraRename3}} function.}

\item{cam_dir_level}{Integer. The level of the directory where the camera name is. This can be a number of levels from the top or bottom directory, specified using the 'from_bottom' argument.}

\item{from_bottom}{Logical. Should cam_dir_level be calculated from the lowest directory (TRUE) or highest directory (FALSE)}
}
\value{
A data frame containing the outpath for the images, the user label, number of pictures, first picture date, and last picture date
}
\description{
This function computes the first and last pictures from the output of the \code{\link{cameraRename3}} function for basic diagnostic purposes.
}
\details{
This function has only been tested to work with the output of \code{\link{cameraRename3}}.
Theoretically it could work with any input that has an outpath, UserLabel, and DateTimeOriginal column but I don't know what it will do.
Update: The above fields seem to be required to ensure that the function does not throw an error. If no UserLabel exists, this function will leave the Label as NA

This function is designed to help you create a CT table by identifying problems.
I would recommend either adjusting your formatting or using this function as a template to build your own.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{cameraRename3}}
}
\concept{camera trapping}
\concept{diagnostics}
\keyword{manip}
