% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{ctdates_fun}
\alias{ctdates_fun}
\title{CT Table date conversion}
\usage{
ctdates_fun(cttable, start.col = 6)
}
\arguments{
\item{cttable}{data frame representing a CT Table.}

\item{start.col}{Defaults to 6. This is the column number in the ct table where dates first appear. It is assumed that all columns after this contain dates.}
}
\value{
The same as the input except with character dates instead of date dates.
}
\description{
This function will convert properly formatted date objects to characters. This is so camtrapR can read the character date. For some reason, the package does not like date-formatted dates.
}
\note{
As with all of my functions, this assumes a very particular formatting for your data. If the CT table is not formatted in this way, then this function will not work. I would recommend either adjusting your formatting or using this function as a template to build your own. These functions are built for very specific purposes and may not generalize well to whatever you need it for. I build them for my own convenience and make no promises that they will work for different situations. As I come across errors, I attempt to further generalize the function but this is done as I go.
}
\seealso{
\code{\link{trapeffort_fun}}
}
\concept{CT tables}
\concept{camera trapping}
\keyword{manip}
