% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{dataOrganize}
\alias{dataOrganize}
\title{DataOrganize}
\usage{
dataOrganize(in.dir, save = F, diagnostics = T)
}
\arguments{
\item{in.dir}{data.frame. The directory containing the camera folders.}

\item{save}{logical. Should the file be saved to the working directory? The default is FALSE. If TRUE, this will create a file called dataorganize.txt containing the output. Be careful, as this can overwrite other files with this name.}

\item{diagnostics}{logical. Should diagnostic information be outputted to the console? This is set to TRUE by default.}
}
\value{
A data frame organized in the same way as the DataOrganize program: Camera Species Year Month Day Hour Minute Second Num_of_Individuals.
}
\description{
This is an R version of the DataOrganize program developed by Jim Sanderson and Grant Harris. While untested, it should provide a little more flexibility in naming of folders than the original DataOrganize program. It also can do basic diagnostics so you can check camera and species names.
}
\examples{
\dontrun{
# No example provided
}
}
\references{
Original DataOrganize program: \url{https://smallcats.org/resources/}
}
\seealso{
\code{\link{APFun_env}}
}
\concept{DataOrganize}
\concept{camera trapping}
\keyword{files}
\keyword{manip}
