% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{deleteEmptyDirs}
\alias{deleteEmptyDirs}
\title{Delete Empty Diretories}
\usage{
deleteEmptyDirs(in.dir, glob = NULL, delete = TRUE)
}
\arguments{
\item{in.dir.}{Where should the function look for directories to delete?}

\item{glob.}{What search term should be used for subsetting? The default is NULL which uses all directories in the in directory}

\item{delete.}{Logical. Defaults to FALSE. Should the function delete empty directories. This cannot be undone to my knowledge.}
}
\value{
A character string containing a list of the empty directories
}
\description{
This function can be used to delete empty directories from the file system.
You can specify a subset or look through all directories in the in directory.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided yet
}

}
\seealso{
\code{\link{cameraRename4}}
}
\concept{directory management}
\keyword{files}
\keyword{manip}
