% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactionsDataOrganize}
\alias{interactionsDataOrganize}
\title{Create an interactions file from a dataorganize file}
\usage{
interactionsDataOrganize(
  do,
  envdata,
  exclude,
  start_date,
  end_date = Sys.Date()
)
}
\arguments{
\item{do}{An AllPictures file produced by DataOrganize.}

\item{envdata}{Environmental variables data frame. This file must have header called "Camera" containing the list of cameras.}

\item{exclude}{species to exclude from the output data frame. Use c() to specify species. If you want keep all items use c(""). Unlike APFun_env, this has no default. Use it to exclude any "species" from the output file.}

\item{start_date}{Start date for the AllPictures file.}

\item{end_date}{End date for pictures. This defaults to the current date.}
}
\value{
A file containing the list of images and species for interactions.
}
\description{
This function uses the old workflow based on DataOrganize and will create an object for use in computing Wildlife Interactions.
}
\details{
This function is nearly identical to the APFun_env except it is simpler because it does not have to calculate an interval for independent events and it outputs a slightly different object.

For a detailed description of what these interactions are, see any of our or UTRGV's reports on the TxDOT camera monitoring projects
}
\examples{
\dontrun{
# No example provided
}
}
\references{
Tewes, M., J. Lombardi, Z. Wardle, and T. Yamashita. 2020. Ocelot and Jaguarundi Monitoring Project: Evaluating the Effectiveness of Wildlife Crossings, Cattle Guards, and Fencing on Road Facilities in Cameron County, Contract No. 57-9XXIA003, 0000018485. Feline Research Program, Caesar Kleberg Wildlife Research Institute, Texas A&M University - Kingsville.

Kline, R., K. Ryer, A. Rivera, T. Yamashita, and T. Hopkins. 2020. Post-construction Monitoring Bi-annual Report for SH 100: May 2019 thru October 2019 (Contract No 57-9XXIA001). The University of Texas Rio Grande Valley.
}
\seealso{
\code{\link{APFun_env}}

\code{\link{interactionsTimelapse}}
}
\concept{DataOrganize}
\concept{interactions}
\keyword{manip}
