% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{mergeFiles}
\alias{mergeFiles}
\title{Merge Files}
\usage{
mergeFiles(in.dir, pattern, save = F)
}
\arguments{
\item{in.dir}{string. The directory containing the files you want to merge}

\item{pattern}{string. The file extension of the files you want to merge. Currently this can only be c(".txt", ".csv", ".xlsx"). The period must be included in the extension name. Note that the ".xlsx" file extension uses the \code{\link[openxlsx]{openxlsx}} package for reading and writing.}

\item{save.}{logical. Should the output be saved to the current working directory? This defaults to FALSE.}
}
\value{
data frame containing the merged files
}
\description{
This function will find all files in a directory with a specified extension and merge them into a single file. It is useful for merge dataorganize or timelapse files for quicker access
}
\details{
This function assumes that you are only merging files that have the same data structure. It does not do any checking of this so please make sure that your data structure is consistent among files before using this function.
This makes it especially useful for merging DataOrganize files or Timelapse outputs.

Currently, this function only allows for the above file extensions. Other extensions could be included in the future. Please let me know if you want a different file extension to be available.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{doFolder}}, \code{\link{doTimelapse}}

\code{\link[openxlsx]{openxlsx}}
}
\concept{DataOrganize}
\concept{camera trapping}
\concept{timelapse}
\keyword{datagen}
\keyword{files}
\keyword{manip}
