% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timelapse.R
\name{movePictures}
\alias{movePictures}
\title{Move pictures from unsorted to sorted folders}
\usage{
movePictures(
  timelapse,
  in.dir,
  out.dir,
  create.dirs,
  type = "none",
  exclude = NULL
)
}
\arguments{
\item{timelapse}{data.frame. A data frame of a timelapse file. Note, this should follow the timelapse template that I typically use}

\item{in.dir}{String. The directory containing the unsorted images}

\item{out.dir}{String. The directory where you want to store the sorted images}

\item{create.dirs}{Logical. Should the function create the directories it needs?}

\item{type}{String. Should you move, copy, or do nothing with the images. Choose one of c('move','copy','none')}

\item{exclude}{String. Which species should not be sorted? The default is NULL which sorts all species. This can take multiple inputs. Use c("Species1", "Species2", "etc") to specify unique species}
}
\value{
list of the full file path to the in files and out files

in.files:
String. Full file paths to the in files

out.files:
String. Full file paths to the out files
}
\description{
This function uses a Timelapse csv file to move or copy images from an unsorted folder to sorted folders based on species and number of individuals (in the same format as required for \code{link{dataorganize}}), although see details.
}
\details{
When this function creates its folder structure, it uses the Individuals column in the Timelapse output.
For some "species" (e.g., ghost, human, bird, rodent), we do not sort these by individual, therefore the Individuals column is a 0.
These species get assigned a folder of 00 for their number of individuals. I do not know how this will affect quality control and the workflow down the line.
Generally, it should not be an issue but could result in NA values in the Individuals column of the \code{\link{APFun_env}} or errors in the \code{\link{dataorganize}} functions in this package.
Once this has been tested, I will update this.
}
\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{dataorganize}}

\code{\link{APFun_env}}
}
\concept{camera trapping}
\concept{timelapse}
\keyword{files}
\keyword{manip}
