% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{prH0}
\alias{prH0}
\title{Calculate true null hypothesis}
\usage{
prH0(p, prior = 0.5, null = TRUE)
}
\arguments{
\item{p}{numeric. The p value from your analysis}

\item{prior}{numeric. The prior probability of a true null hypothesis.
0.5 indicates an uninformative prior.}

\item{null}{logical. Do you want to test whether the null hypothesis is true or the alternative?}
}
\value{
value. The requested probability given the p value and prior.
}
\description{
This function calculates the probability of a true null hypothesis
}
\examples{
\dontrun{
# No example provided
}
}
\references{
Sellke, T., M. J. Bayarri, and J. O. Berger. 2001. Calibration of ρ Values for Testing Precise Null Hypotheses. The American Statistician 55:62-71.

Benjamin, D. J., and J. O. Berger. 2019. Three Recommendations for Improving the Use of p-Values. The American Statistician 73:186-191.
}
\concept{p value}
\concept{probability}
\keyword{arith}
\keyword{htest}
\keyword{utilities}
