% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{subsetImages}
\alias{subsetImages}
\title{Subset Images}
\usage{
subsetImages(
  in.dir,
  out.dir,
  ext = c(".jpg", ".mp4"),
  datatype,
  from,
  to,
  date.col,
  species.col,
  create.dirs = F,
  type = "none"
)
}
\arguments{
\item{in.dir}{String. The directory containing the images folders. This must the folder containing either the raw or sorted camera folders.}

\item{out.dir}{String. The directory where you want to move or copy your subsetted images.}

\item{ext}{String. Defaults to c(".jpg", ".mp4"). What file extensions should the function look for for subsetting?}

\item{datatype}{String. Does the in.dir contain raw or sorted images. This can only be c("raw", "sorted").}

\item{from}{String. Optional. The start date for images you want to subset. This is not required. See details below.}

\item{to}{String. Optional. The end date for images you want to subset. This is not required. See details below.}

\item{date.col}{String. Optional. Which date collected folders should be included in the subset. This is not required. See details below. This cannot be specified when datatype = "sorted".}

\item{species.col}{String. Optional. Which species folders should be included in the subset. This is not required. See details below. This cannot be specified when datatype = "raw".}

\item{create.dirs.}{Logical. Defaults to FALSE. Should the function create the directories it needs to do a file transfer?}

\item{type.}{String. Should you move, copy, or do nothing with the images. Choose one of c('move','copy','none').}
}
\value{
data.frame containing the old file path and new file path
}
\description{
Subset images from a larger image directory
}
\details{
When this function does its subset, you must specify at least one of from, to, date.col, or species.col.
date.col can only be used with raw data and species.col can only be used with sorted data.
When more than one of these is included, the subset will be created from all parameters provided.
The function will tell you what it is using for the subset.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{unsortImages}}, \code{\link{movePictures}}, \code{\link{removeGhosts}}
}
\concept{camera trapping}
\concept{subsetting images}
\keyword{files}
\keyword{manip}
