% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisprep.R
\name{summarizeEvents}
\alias{summarizeEvents}
\title{Summarize camera trap events, including setting up for regression analyses, occupancy models, n-mixture models}
\usage{
summarizeEvents(
  x,
  ct,
  unit,
  include,
  camOP,
  out_form = "all",
  out_data = "all",
  out_correction = "all"
)
}
\arguments{
\item{x}{A data frame produced by the \code{\link{calculateEvents}} function available in this package.}

\item{ct}{A CT Table following the format from the package camtrapR.}

\item{unit}{The temporal unit used to summarize the data. This should a unit >= 1 day. Weeks, months, or years can all be called using this argument.
See details for how this function handles summarizing units greater than days.}

\item{include}{String of characters. This should be the names of the species that you want to include in this analysis.}

\item{camOP}{Arguments passed to \code{\link[camtraR]{cameraOperation}}. The most important ones of these are
stationCol, setupCol, and retrievalCol. hasProblems is often used as well.
If multiple cameras per site, you must specify cameraCol, byCamera, allCamsOn, and camerasIndepedent.
See \code{\link[camtrapR]{cameraOperation}} for details.}

\item{out_form}{Should the data output in long format, wide format or both. Acceptable inputs are c("all", "l", "long", "w", "wide"). The maximum length of this input is 2. Do not include "all" if providing multiple arguments.}

\item{out_data}{Should the data output number of detections (DE), abundance (AB; number of individuals), or presence/absence (PA). Acceptable inputs are c("all", "DE", "detections", "AB", "abundance", "PA", "presence"). The maximum length of this input is 3. Do not include "all" if providing multiple arguments.}

\item{out_correction}{Should the data apply no correction for active camera trap nights (raw, none; events),
remove all data when camera traps were inactive for more than half the time interval (rm, rmInactive; events),
calculate a number of detections/abundance per day for each time interval (pd, perDay; events/activenights),
or calculate an adjusted number of detections/abundance for each time interval (pu, perUnit; events/activenights * totalnights).
Acceptable inputs are c("all", "none", "raw", "rm", "rmInactive", "pd", "perDay", "pu", "perUnit").
The maximum length of this input is 4. Do not include "all" if providing multiple arguments.}
}
\value{
Usually a 2-leveled list with species (chosen using the include argument) as the top level and requested outputs (chosen using out_form, out_data, and out_correction) below it.
However, this function will coerce any list that is length 1 to a data.frame, simplifying the return.
}
\description{
Summarize raw camera trap events by a user-specified time period for various analyses
}
\details{
Using units other than days: If you use units other than days, the function will start the time interval on the first day of the period.
For example, if "1 month" is specified as the unit, the function will start each period on the 1st of the month.
If "1 week" is specified as the unit, the function will start each period on Sunday.
If "1 year" is specified as the unit, the function will start on the January 1.

This function deprecates the \code{\link{occFun}} function. All capabilities from that function are implemented here in a simpler, faster, and more versatile way. Avoid using occFun in favor of this function in the future.

A note about the outputs: This function will output up to 18 items (2 forms, 3 datas, and 3 corrections) and does a combination of each. Currently it is not possible to request specific outputs of multiple forms, datas, and corrections.
}
\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{calculateEvents}} \code{\link{occFun}}
}
