% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.R
\name{timeConvert}
\alias{timeConvert}
\title{Convert Time Zones from GPS collar data}
\usage{
timeConvert(ds, date.col, time.col, LocalTZ)
}
\arguments{
\item{ds}{The original GPS collar data.}

\item{date.col}{The column containing the original dates that need to be converted.}

\item{time.col}{The column containing the original times that need to be converted.}

\item{LocalTZ}{The time zone that you want to convert to. This must an R-recognized string for a given time zone. This function utilizes the with_tz function from the lubridate package. See the help for that package for details.}
}
\value{
Your original data frame with the new date and time columns appended to the end as LocalDate and LocalTime
}
\description{
This function converts date-time information from a GPS collar from UTC time to your local time zone. This should be used when GPS collar data is not already converted and you need to keep separate "date" and "time" columns in your data.
}
\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link[lubridate]{with_tz}}
}
\concept{collars}
\keyword{manip}
