% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{trapEffort}
\alias{trapEffort}
\title{Camera trapping trap night effort}
\usage{
trapEffort(ct, camOP)
}
\arguments{
\item{ct}{data.frame. A data frame formatted as a CT table for the  \code{\link[camtrapR]{camtrapR-package}}}

\item{camOP}{list. Arguments passed to \code{\link[camtraR]{cameraOperation}}. The most important ones of these are
stationCol, setupCol, and retrievalCol. hasProblems is often used as well.
If multiple cameras per site, you must specify cameraCol, byCamera, allCamsOn, and camerasIndepedent.
See \code{\link[camtrapR]{cameraOperation}} for details.}
}
\value{
A data frame containing the name of the Station (and possibly the session number and camera ID, depending on the inputs of camOP),
active camera trap nights (activenights), and total camera trap nights (totalnights).
}
\description{
This function calculates the number of active camera trap nights and total camera trap nights using an inputted CT table, formatted based on camtrapR specifications.Required columns are setup date and retrieval date and theoretically, the table should have problems. I have never tested it on a dataset without any problems.
}
\details{
Make sure that your CT table is formatted properly. See the \code{\link[camtrapR]{camtrapR-package}} documentation for details. That is the only way this function works. Also, at some point camptrapR had removed its support for dates in "Date" or "POSIXct" format so dates had to be in character format. You can use my ctdates_fun function to fix this in a CT table. This may not be the case anymore and they may have fixed this issue.

A CT table formatted for \code{\link[camtrapR]{camtrapR}} is required. This CT table is used in other functions that utilize the camtrapR package and must be formatted properly.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{ctDates}} \code{\link[camtrapR]{cameraOperation}}
}
\concept{camera trapping}
\keyword{manip}
