% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycontrol.R
\name{unsortImages}
\alias{unsortImages}
\title{Move pictures from sorted to unsorted folders}
\usage{
unsortImages(in.dir, out.dir, date.col, type, create.dirs = T)
}
\arguments{
\item{in.dir}{string. The directory containing the sorted camera folders. This can have lengths greater than 1}

\item{out.dir}{string. The directory where you want to unsort files to. This can have length 1 or equal to the in.dir}

\item{type}{String. Should you move, copy, or do nothing with the images. Choose one of c('move','copy','none')}

\item{create.dirs}{Logical. Should the function create the directories it needs?}

\item{datecol}{string. The date collected of the batch/es of pictures that need to be unsorted. This must be the same length as the in.dir}
}
\value{
list of the full file path to the in files and out files

in.files:
String. Full file paths to the in files

out.files:
String. Full file paths to the out files
}
\description{
This function is designed to take sorted images and return them to Raw data structure format. It is the reverse process to the movePictures function
}
\details{
This function assumes that the sorted pictures are sorted in the same way as what is created by the movePictures function.
Therefore, it can be considered a reverse procedure to the movePictures function. This can be used to create new backups of raw data if any are lost or if the raw data is otherwise unavailable.

This function has not gone through full testing and may have problems. Please report issues as they come up.
}
\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{movePictures}}
}
\concept{camera trapping}
\concept{sorting}
\keyword{files}
\keyword{manip}
