% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{bestPics}
\alias{bestPics}
\title{Extract Best Pictures from Timelapse}
\usage{
bestPics(timelapse, in.dir, out.dir, copy = T, sorted = F)
}
\arguments{
\item{timelapse}{A data frame from a timelapse file}

\item{in.dir}{String. The directory containing the pictures. This can either be an unsorted directory or a sorted directory. Keep in mind, these assume my very specific directory format.}

\item{out.dir}{String. The directory where you want the best pictures to be stored. This can be anything.}

\item{copy}{Logical. Should the images be copied into the new directory?}

\item{sorted}{Logical. Is the in.dir an unsorted folder or a sorted folder?}
}
\value{
A data frame containing the file path info of the best pics.
}
\description{
This function uses a timelapse csv to extract photos tagged as a best picture and copies them to a specified location.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{movePictures}}
}
\concept{camera trapping}
\concept{timelapse}
\keyword{files}
\keyword{manip}
