% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisprep.R
\name{calculateEvents}
\alias{calculateEvents}
\title{Calculate number of events from a file organized like a dataOrganize file.}
\usage{
calculateEvents(
  do,
  envdata,
  sort.col = "Camera",
  exclude = c("ghost"),
  start_date,
  end_date = Sys.Date(),
  interval = NULL,
  all.pics = FALSE
)
}
\arguments{
\item{do}{data frame. A DataOrganize file produced by the DataOrganize program or the dataOrganize function in this package.}

\item{envdata}{Environmental variables data frame. This file must have header called "Camera" containing the list of cameras.}

\item{sort.col}{string. Column that you want to sort pictures by to create independent events. This should be one of c("Camera", "Site", "Station") depending on your wording for sites. This defaults to "Camera".}

\item{exclude}{string. species to exclude from the output data frame. Use c() to specify species. This defaults to excluding ghosts. If you want keep all items use c("").}

\item{start_date}{string. Start date for the AllPictures file. If you want all data, set this to an arbitrarily early date.}

\item{end_date}{string. End date for pictures. This defaults to the current date.}

\item{interval}{integer. Time in seconds between pictures for an independent event. See Details for some normal interval times. This is only necessary for the main function. It defaults to no interval.}

\item{all.pics}{Logical. Whether or not you want to return all the pictures or just independent events. If all.pics = TRUE, all pictures will be returned with an index number associated with independent events. This defaults to FALSE.}
}
\value{
A data frame of camera data. Output varies depending on whether or not all.pics is true or false.

all.pics = FALSE:
Returns a data frame of only independent events at the specified interval.

all.pics = TRUE:
Returns a data frame of all pictures from the input file. A index number is added associated with independent events.
}
\description{
Function for calculating the number of events from raw camera trap images and adding camera/site-specific information to the output.
}
\details{
Suggested interval times: 61 = 1 minute, 1801 = 30 minutes. 3601 = 1 hour

One second should be added to the total to ensure the full interval is included.

When interval is set to NULL, the function will use an interval of 1 second and therefore does not summarize any events.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{doTimelapse}}, \code{\link{doFolder}}

\code{\link{actFun}}, \code{\link{summarizeEvents}}
}
\concept{DataOrganize}
\concept{camera trapping}
\keyword{datagen}
