% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganize.R
\name{dataorganize}
\alias{dataorganize}
\title{DataOrganize}
\usage{
dataorganize(inputdir, diagnostics = T)
}
\arguments{
\item{inputdir}{The directory containing the camera folders.}

\item{diagnostics}{Logical. Should diagnostic information be outputted to the console? This is set to TRUE by default.}
}
\value{
A data frame organized in the same way as the DataOrganize program: Camera Species Year Month Day Hour Minute Second Num_of_Individuals.
}
\description{
This is an R version of the DataOrganize program developed by Jim Sanderson and Grant Harris. While untested, it should provide a little more flexibility in naming of folders than the original DataOrganize program. It also can do basic diagnostics so you can check camera and species names.
}
\references{
Original DataOrganize program: \url{https://smallcats.org/resources/}
}
\seealso{
\code{\link{APFun_env}}
}
\concept{DataOrganize}
\concept{camera trapping}
\keyword{files}
\keyword{manip}
