% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactions.R
\name{interactionsTimelapse}
\alias{interactionsTimelapse}
\title{Create an Interactions File from a Timelapse Output}
\usage{
interactionsTimelapse(
  images,
  envdata,
  exclude,
  create.dirs = T,
  copy.files = T
)
}
\arguments{
\item{images}{A timelapse csv, formatted using our Timelapse template.}

\item{envdata}{Environmental variables data frame. This file must have header called "Camera" containing the list of cameras, "Site" containing the site name, "Side" indicating which side the camera is on, and "Type" indicating what type of structure the camera is at.}

\item{exclude}{species to exclude from the output data frame. Use c() to specify species. If you want keep all items use c(""). Unlike APFun_env, this has no default. Use it to exclude any "species" from the output file.}

\item{create.dirs}{Logical. Defaults to TRUE. Should new directories be checked for and created by R if necessary?}

\item{copy.files}{Logical. Defaults to TRUE. Should image files be copied to the appropriate directories?}
}
\value{
A list containing the data file and a file of the old and new paths and names.

Interactions:
The Interactions data file.

Files:
File used for copying files.
}
\description{
This function copies images for interactions to a new folder and creates an interactions file.
}
\details{
NOTE: R has no good way of handling duplicate file names so it will skip duplicates when copying. Depending on how images are named, this could be an issue. The easiest way to check is to check the number of images in the main folder.
}
\examples{
\dontrun{
# No example provided
}
}
\references{
Tewes, M., J. Lombardi, Z. Wardle, and T. Yamashita. 2020. Ocelot and Jaguarundi Monitoring Project: Evaluating the Effectiveness of Wildlife Crossings, Cattle Guards, and Fencing on Road Facilities in Cameron County, Contract No. 57-9XXIA003, 0000018485. Feline Research Program, Caesar Kleberg Wildlife Research Institute, Texas A&M University - Kingsville.

Kline, R., K. Ryer, A. Rivera, T. Yamashita, and T. Hopkins. 2020. Post-construction Monitoring Bi-annual Report for SH 100: May 2019 thru October 2019 (Contract No 57-9XXIA001). The University of Texas Rio Grande Valley.
}
\seealso{
\code{\link{interactionsDataOrganize}}
}
\concept{interactions}
\concept{timelapse}
\keyword{manip}
