% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{movePictures}
\alias{movePictures}
\title{Move pictures from unsorted to sorted folders}
\usage{
movePictures(
  timelapse = NULL,
  do = NULL,
  in.dir,
  out.dir,
  create.dirs,
  type = "none",
  exclude = NULL,
  ...
)
}
\arguments{
\item{timelapse}{data.frame. A data frame of a timelapse file. Note, this should follow the timelapse template that I typically use. You cannot specify both a timelapse file and a Dataorganize file using this function}

\item{do}{data.frame. A data frame of a DataOrganze file. This can be either from the DataOrganize program or from the \code{\link{dataOrganize}} function in this package. You cannot specify both a timelapse and DataOrganize file using this function.}

\item{in.dir}{String. The directory containing the root folder for the timelapse file. For example, if all your images were a folder called "images" which sits in an external drive, labelled "F:", then you would specify the in.dir as "F:".}

\item{out.dir}{String. The directory where you want to store the sorted images}

\item{create.dirs}{Logical. Should the function create the directories it needs?}

\item{type}{String. Should you move, copy, or do nothing with the images. Choose one of c('move','copy','none')}

\item{exclude}{String. Which species should not be sorted? The default is NULL which sorts all species. This can take multiple inputs. Use c("Species1", "Species2", "etc") to specify unique species}

\item{...}{Additional arguments used when specifying a DataOrganize file. Only img_format and do_format are used. All other inputs are ignored. Both img_format and do_format can only be c("serial", "original") and are used to indicate whether a serial number has been included in the image names and/or DataOrganize file.}
}
\value{
list of the full file path to the in files and out files

in.files:
String. Full file paths to the in files

out.files:
String. Full file paths to the out files
}
\description{
This function uses a Timelapse csv file to move or copy images from an unsorted folder to sorted folders based on species and number of individuals (in the same format as required for \code{link{dataorganize}}), although see details.
}
\details{
When this function creates its folder structure, it uses the Individuals column in the Timelapse output.
For some "species" (e.g., ghost, human, bird, rodent), we do not sort these by individual, therefore the Individuals column is a 0.
These species get assigned a folder of 00 for their number of individuals. I do not know how this will affect quality control and the workflow down the line.
Generally, it should not be an issue but could result in NA values in the Individuals column of the \code{\link{APFun_env}} or errors in the \code{\link{dataorganize}} functions in this package.
Once this has been tested, I will update this.

This function can "sort" pictures using either a timelapse csv or a dataorganize text file.
The inclusion of a dataorganize file was to better accommodate data sorted manually (i.e., not using Timelapse) and its primary use is when sorted images need to be replaced (e.g., the images got corrupted or lost somehow).
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
If you want to make these functions work, I would recommend either adjusting your formatting or using this function as a template to build your own.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{dataOrganize}}

\code{\link{APFun_env}}
}
\concept{camera trapping}
\concept{timelapse}
\keyword{files}
\keyword{manip}
