% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CameraAnalysis.R
\name{occFun}
\alias{occFun}
\title{Occupancy Analysis Data Setup}
\usage{
occFun(
  x,
  ct,
  unit,
  subset,
  stationCol,
  sessionCol = NULL,
  ct_probs = T,
  count = F
)
}
\arguments{
\item{x}{A data frame produced by the APFun_env function available in this package.}

\item{ct}{A CT Table following the format from the package camtrapR.}

\item{unit}{The temporal unit for dividing up your data. This must be in units "days".}

\item{subset}{A character vector indicating which species to prepare data for.}

\item{stationCol}{String. A character string for the column used to identify sites.}

\item{sessionCol}{Character. A character string for the column used to identify sessions. This is currently untested. Use with caution.}

\item{ct_probs}{Logical. Does the CT table have problems?}

\item{count}{Logical. Should presence/absence (FALSE) or counts (TRUE) be outputted?}
}
\value{
A LIST object by species chosen in the subset input with a table of sites (rows) and dates (columns).
}
\description{
This function modifies data for use in occupancy modelling using Unmarked or a separate program, I think.
}
\details{
Please note that I personally have never tested the output of this function with an occupancy model although I am told that it works. Please use this with caution and test it properly before using it for any real research.
}
\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{APFun_env}}
}
\concept{camera trapping}
\concept{occupancy analysis}
\keyword{datagen}
\keyword{manip}
