% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitycontrol.R
\name{trapEffort}
\alias{trapEffort}
\title{Camera trapping effort (Trap nights)}
\usage{
trapEffort(cttable, group, sessions = F, sessioncol = NULL)
}
\arguments{
\item{cttable}{data.frame. A data frame formatted as a CT table.}

\item{group}{String. A column in the CT Table indicating how camera trap nights should be calculated. Generally accepted are c("Camera", "Site", "Station").}

\item{sessions}{Logical. Does your data have multiple sessions (e.g., field seasons, etc.)? This defaults to FALSE.}

\item{sessioncol}{String. What column distinguishes sessions? This defaults to NULL. This is only needed if sessions is set to T.}
}
\value{
A data frame containing the items from the group column, session column (if included), active camera trap nights, and total camera trap nights.
}
\description{
This function calculates the number of active camera trap nights and total camera trap nights using an inputted CT table, formatted based on camtrapR specifications.Required columns are setup date and retrieval date and theoretically, the table should have problems. I have never tested it on a dataset without any problems.
}
\details{
Make sure that your CT table is formatted properly. See the \code{\link[camtrapR]{camtrapR-package}} documentation for details. That is the only way this function works. Also, at some point camptrapR had removed its support for dates in "Date" or "POSIXct" format so dates had to be in character format. You can use my ctdates_fun function to fix this in a CT table. This may not be the case anymore and they may have fixed this issue.

A CT table formatted for \code{\link[camtrapR]{camtrapR}} is required. This CT table is used in other functions that utilize the camtrapR package and must be formatted properly.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{ctDates}} \code{\link[camtrapR]{cameraOperation}}
}
\concept{camera trapping}
\keyword{manip}
