% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{unsortImages}
\alias{unsortImages}
\title{Move pictures from sorted to unsorted folders}
\usage{
unsortImages(in.dir, out.dir, date.col, type, create.dirs = T)
}
\arguments{
\item{in.dir}{string. The directory containing the sorted camera folders. This can have lengths greater than 1}

\item{out.dir}{string. The directory where you want to unsort files to. This can have length 1 or equal to the in.dir}

\item{date.col}{string. The date collected of the batch/es of pictures that need to be unsorted. This must be the same length as the in.dir}

\item{type}{String. Should you move, copy, or do nothing with the images. Choose one of c('move','copy','none')}

\item{create.dirs}{Logical. Should the function create the directories it needs?}
}
\value{
list of data frames of each each in.dir directory. Each data frame contains the following information:

in.files:
String. Full file paths to the in files

out.files:
String. Full file paths to the out files

duplicate:
Logical. Is the file a duplicate file
}
\description{
This function is designed to take sorted images and return them to Raw data structure format. It is the reverse process to the movePictures function
}
\details{
This function assumes that the sorted pictures are sorted in the same way as what is created by the movePictures function.
Therefore, it can be considered a reverse procedure to the movePictures function. This can be used to create new backups of raw data if any are lost or if the raw data is otherwise unavailable.

This function will only copy the first instance of an image and will ignore duplicates. This is because duplicate images can arise in sorted images when more than one species is detected in a single image.

This function has not gone through full testing and may have problems. Please report issues as they come up.
}
\section{{Standard Disclaimer}}{
 As with most of the functions in this package, using this function assumes that you have been following my normal workflow, including the particular formatting that these functions assume.
These functions are built for very specific purposes and may not generalize well to whatever you need it for.
I build them for my own use and make no promises that they will work for different data formatting situations.
As I come across errors, I attempt to further generalize the functions but this is done as I go.
}

\examples{
\dontrun{
# No example provided
}
}
\seealso{
\code{\link{movePictures}}
}
\concept{camera trapping}
\concept{sorting}
\keyword{files}
\keyword{manip}
